Cleanup

+++ Src/Device/CUDAContext.h
@@ -4,13 +4,13 @@
 namespace CUDAContext {
 	inline int compute_capability = -1;
 
-	inline unsigned long long total_memory;
+	inline size_t total_memory;
 
 	// Creates a new CUDA Context
 	void init();
 	void free();
 
-	unsigned long long get_available_memory(); // Available memory on GPU in bytes
+	size_t get_available_memory(); // Available memory on GPU in bytes
 
 	unsigned get_shared_memory(); // Available shared memory in bytes (per Block)
